[SPARK-39273][PS][TESTS] Make PandasOnSparkTestCase inherit ReusedSQLTestCase
### What changes were proposed in this pull request?

This PR proposes to make `PandasOnSparkTestCase` inherit `ReusedSQLTestCase`.

### Why are the changes needed?

We don't need this:

```python
    classmethod
    def tearDownClass(cls):
        # We don't stop Spark session to reuse across all tests.
        # The Spark session will be started and stopped at PyTest session level.
        # Please see pyspark/pandas/conftest.py.
        pass
```

anymore in Apache Spark. This has existed to speed up the tests when the codes are in Koalas repository where the tests run sequentially in single process.

In Apache Spark, we run in multiple processes, and we don't need this anymore.

### Does this PR introduce _any_ user-facing change?

No, test-only.

### How was this patch tested?

Existing CI should test it out.

Closes #36652 from HyukjinKwon/SPARK-39273.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>