[SPARK-39296][CORE][SQL] Replcace `Array.toString` with `Array.mkString`
### What changes were proposed in this pull request?
Some code in Spark as follows:

https://github.com/apache/spark/blob/46662c42824f85578c95921435f7cea6dc625b2d/core/src/main/scala/org/apache/spark/deploy/worker/Worker.scala#L517-L520

`appDirs` is an `Array` and `Array.toString` will print as `className#hashCode`, this result seems meaningless, this PR replace `Array.toString` with `Array.mkString` in similar scenario to make the printing of the `Array` meaningful.

### Why are the changes needed?
Replace `Array.toString` with `Array.mkString` to make the printing of the `Array` meaningful.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Pass Github Actions

Closes #36677 from LuciferYang/array-mkstring.

Authored-by: yangjie01 <yangjie01@baidu.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>