[SPARK-39267][SQL] Clean up dsl unnecessary symbol
### What changes were proposed in this pull request?

add deprecated tag for two methods in dsl:
- subquery(alias: Symbol)
- as(alias: Symbol)

### Why are the changes needed?

dsl is a test helper file which provide easy used functions. But some of these are unnecessary, for example:
`def subquery(alias: Symbol): LogicalPlan`
For a subquery, we only need the name, so a string type parameter is enough.

Besides, the symbol is not a very stable interface in scala. It should be better to clean up it if unnecessary.

### Does this PR introduce _any_ user-facing change?

no

### How was this patch tested?

pass exists test

Closes #36646 from ulysses-you/clenup-dsl.

Authored-by: ulysses-you <ulyssesyou18@gmail.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>