[SPARK-30835][SPARK-39018][CORE][YARN] Add support for YARN decommissioning when ESS is disabled
### What changes were proposed in this pull request?

This PR tries to make spark leverage the knowledge of the node loss in future, and tries to adjust the scheduling of tasks to minimise the impact on the application.

This PR Adds support for YARN decommissioning & pre-emption

Prevent new tasks from starting on any executors on decommissioning Nodes. This PR only supports Yarn Decommissioning when External Shuffle Service is Disabled.  There is a separate sub task to support Yarn Decommissioning when Shuffle Service is enabled. [SPARK-39024](https://issues.apache.org/jira/browse/SPARK-39024)

### Why are the changes needed?
Add the support to handle the Node Decommissioning for Yarn cluster manger in Spark

### Does this PR introduce any user-facing change?
NO

### How was this patch tested?
Added the Unit Test and run the manual test.

Closes #35683 from abhishekd0907/yarn-node-decom.

Lead-authored-by: Abhishek Dixit <abhi4.dixit8@gmail.com>
Co-authored-by: abhishekd0907 <43843989+abhishekd0907@users.noreply.github.com>
Signed-off-by: Mridul Muralidharan <mridul<at>gmail.com>