[SPARK-39394][SPARK-39253][DOCS][FOLLOW-UP] Fix the PySpark API reference links in the documentation
### What changes were proposed in this pull request?

This PR is a followup of https://github.com/apache/spark/pull/36782 and https://github.com/apache/spark/pull/36647. Both PRs missed changing the URLs in other places in the documentation.

### Why are the changes needed?

To fix the broken links so users can read the correct documentation.

### Does this PR introduce _any_ user-facing change?

Yes, it fixes the broken links in the Spark documentation.

### How was this patch tested?

Manually tested.

Closes #36997 from HyukjinKwon/SPARK-39394-SPARK-39253.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>