Revert [SPARK-35223] Add IssueNavigationLink
### What changes were proposed in this pull request?

IDEA may re-generate the config files and it's annoying to manually exclude it from commit as the IDEA config file is under version control.

This PR removes the IDEA config files and ignores it in git. This is basically a revert of https://github.com/apache/spark/pull/32337

We can update the developer guideline about how to configure IDEA to show Github/JIRA links.

### Why are the changes needed?

avoid git troubles.

### Does this PR introduce _any_ user-facing change?

no

### How was this patch tested?

N/A

Closes #37046 from cloud-fan/tiny.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>