[MINOR][SQL][DOC] Fix doc for `WindowSpec.rowsBetween`
### What changes were proposed in this pull request?

Fix Doc for `WindowSpec.rowsBetween`

### Why are the changes needed?

Given a row based sliding frame with a lower bound offset of -1 and a upper bound offset of +2.

The frame for row with index 5 should range from index 4 to index 7. But current doc says it would range from index 4 to index 6.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Doc changes only.

Closes #37061 from iRakson/windowDocFix.

Authored-by: iRakson <raksonrakesh@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>