[SPARK-39675][SQL] Switch 'spark.sql.codegen.factoryMode' configuration from testing purpose to internal purpose
### What changes were proposed in this pull request?

This PR proposes to switch 'spark.sql.codegen.factoryMode' configuration from testing purpose to internal purpose by removing `Utils.isTesting`.

In addition, it changes the documentation in 'spark.sql.codegen.factoryMode' to say that this is only for internal purpose, and NOT supposed to be set by end users.

### Why are the changes needed?

We can actually leverage this configuration in other testing beds that mimic production environment to make sure the fallback codegen works in production too.

Another point is about benchmarking in the production enviornment.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Manual testing.

Closes #37076 from HyukjinKwon/SPARK-39675.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>