[SPARK-39693][INFRA] Do Not Execute tpcds-1g-gen for Benchmarks Other Than TPCDSQueryBenchmark
### What changes were proposed in this pull request?
Currently `tpcds-1g-gen` runs for any benchmarks even that do not require TPC-DS data on Github Actions.

This PR proposes to skip running `tpcds-1g-gen` if the benchmark class does not contain `TPCDSQueryBenchmark` or `*` based on the discussion on #37020

### Why are the changes needed?
This PR should save time to launch benchmarks on Github Actions

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Tested on Github Actions.

Closes #37120 from kazuyukitanimura/SPARK-39693.

Authored-by: Kazuyuki Tanimura <ktanimura@apple.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>