[SPARK-39767][SQL] Remove UnresolvedDBObjectName and add UnresolvedIdentifier
### What changes were proposed in this pull request?

This PR removes `UnresolvedDBObjectName` and adds a new `UnresolvedIdentifier` to replace it. The same for the resolved plans. The motivation is, `UnresolvedDBObjectName` has a `isNamespace` flag. If it's true, then `UnresolvedDBObjectName` has no difference from `UnresolvedNamespace`. It's a bit weird to have 2 options to do the same thing. If we need to resolve a namespace, we should always use `UnresolvedNamespace`.

### Why are the changes needed?

code simplification

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

existing tests

Closes #37178 from cloud-fan/refactor.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>