[SPARK-39625][SPARK-38904][SQL] Add Dataset.as(StructType)
### What changes were proposed in this pull request?

This PR adds a new API `Dataset.as(StructType)` to "cast" the entire dataframe to the specified schema. Its behavior is very similar to table insertion where we need to adjust the input query to match the table schema, but it's extended to work for inner fields as well:
* Reorder columns and/or inner fields to match the specified schema
* Project away columns and/or inner fields that are not needed by the specified schema
* Cast the columns and/or inner fields to match the data types in the specified schema

### Why are the changes needed?

For lakehouse tables, schema evolution is a common feature. It's more robust to specify the expected schema at the beginning so that the data pipeline can always be run with the same input schema.

### Does this PR introduce _any_ user-facing change?

Yes, it adds a new API.

### How was this patch tested?

a new test suite

Closes #37011 from cloud-fan/as.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>