[SPARK-39469][SQL] Infer date type for CSV schema inference
### What changes were proposed in this pull request?

1. Add a new `inferDate` option to CSV Options. The description is:
> Whether or not to infer columns that satisfy the `dateFormat` option as `Date`. Requires `inferSchema` to be true. When `false`, columns with dates will be inferred as `String` (or as `Timestamp` if it fits the `timestampFormat`) Legacy date formats in `Timestamp` columns cannot be parsed with this option.

An error will be thrown if `inferDate` is true when SQL Configuration LegacyTimeParserPolicy is `LEGACY`. This is to avoid incorrect schema inferences from legacy time parsers not doing strict parsing.

The `inferDate` option should prevent performance degradation for users who don't opt-in.

2. Modify InferField in CSVInferSchema.scala to include Date type.

If `typeSoFar` in `inferField` is Date, Timestamp or TimstampNTZ, we will first attempt to parse Date and then parse Timestamp/TimestampNTZ. The reason why we attempt to parse date for `typeSoFar`=Timestamp/TimestampNTZ is because of the case where a column contains a timestamp entry and then a date entry - we should detect both of the data types and infer the column as a timestamp type.

Example:
```
Seq("2010|10|10", "2010_10_10")
  .toDF.repartition(1).write.mode("overwrite").text("/tmp/foo")
spark.read
  .option("inferSchema", "true")
  .option("header", "false")
  .option("dateFormat", "yyyy|MM|dd")
  .option("timestampFormat", "yyyy_MM_dd").csv("/tmp/foo").printSchema()
```
Result:
```
root
 |-- _c0: timestamp (nullable = true)
```

3. Also modified `makeConverter` in `UnivocityParser` to handle Date type entries in a timestamp type column to properly parse the above example.

### Does this PR introduce _any_ user-facing change?

The new behavior of schema inference when `inferDate = true`:
1. If a column contains only dates, it should be of “date” type in the inferred schema
--> If the date format and the timestamp format are identical (e.g. both are yyyy/mm/dd), entries will default to being interpreted as Date
3. If a column contains dates and timestamps, it should be of “timestamp” type in the inferred schema

### How was this patch tested?

Unit tests were added to `CSVInferSchemaSuite` and `UnivocityParserSuite`. An end to end test is added to `CSVSuite`

### Benchmarks:

`inferDate` increases parsing/inference time in general. The impact scales with the number of rows (and not the number of columns). For columns of date type (which would be inferred as timestamp when `inferDate=false`), inference and parsing takes 30% longer. The performance impact is much greater on columns of timestamp type (taking 30x longer than `inferDate=false`) - due to trying each timestamp as a date (and throwing an error) during the inference step.

#### Number of seconds taken to parse each CSV file with `inferDate true` and `inferDate false`

|                                             | inferDate=False | inferDate=True | master branch |
|---------------------------------------------|-----------------|----------------|---------------|
| Small file (<100 row/col). Mixed data types | 0.32            | 0.33           |                |
| 100K rows. 4 columns. Mixed data types.     | 0.70            | 2.80           | 0.70          |
| 20k columns. 4 rows. Mixed Data types.      | 16.32           | 15.90          | 13.5           |
| Large file. Only date type.                 | 2.15            | 3.70           | 2.10          |
| Large file. Only timestamp type.            | 2.60            | 77.00          | 2.30          |

Results are the average of 3 trials with the same machine.

Over multiple runs, master branch benchmark times have also shown results that are slower than `inferDate=false` (although the average is slightly faster). Given the +/- 20% variance in results between trials, master branch benchmark results are roughly similar to `inferDate=False` results.

Closes #36871 from Jonathancui123/SPARK-39469-date-infer.

Lead-authored-by: Jonathan Cui <jonathan.cui@databricks.com>
Co-authored-by: Jonathan Cui <jonathancui123@gmail.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>