[SPARK-37888][SQL][TESTS][FOLLOWUP] Don't check the `Created By` field in `DESCRIBE TABLE` tests
### What changes were proposed in this pull request?
In the PR, I propose to do not check the field `Created By` in tests that check output of the `DESCRIBE TABLE` command.

### Why are the changes needed?
The field `Created By` depends on the current Spark version, for instance `Spark 3.4.0-SNAPSHOT`. Apparently, the tests that check the field depend on Spark version. The changes are needed to avoid dependency from Spark version, and to don't change the tests when bumping Spark version.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
By running the modified tests:
```
$ build/sbt -Phive-2.3 -Phive-thriftserver "test:testOnly *DescribeTableSuite"
```

Closes #37299 from MaxGekk/unify-describe-table-tests-followup.

Authored-by: Max Gekk <max.gekk@gmail.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>